feat: changes

+++ controller/ctfManager.js
@@ -83,22 +83,31 @@ async function createContainer(imageName, userId, flag = Math.random().toString(
  */
 async function deleteAllContainers() {
     // TODO: IN PROCESS
-    console.log('Deleting all containers');
-    try {
-        const container = await getContainerLabel(process.env.CTF_LABEL, true);
-        if (container == null) {
-            console.log(`The container ${containerName} does not exist.`);
-        } else {
-            await container.kill();
-            await container.remove();
-            console.log(`Container ${containerName} deleted successfully.`);
-        }
-    } catch (error) {
-        console.error('ERROR: deleting container', error);
-        throw error;
-    }
+    // console.log('Deleting all containers');
+    // try {
+    //     const container = await getContainerLabel("ctf", true);
+    //     if (container == null) {
+    //         console.log(`No containers found related with ctf.`);
+    //     } else {
+    //         await container.kill();
+    //         await container.remove();
+    //     }
+    //     console.log(`CTF containers deleted successfully.`);
+    // } catch (error) {
+    //     console.error('ERROR: deleting container', error);
+    //     throw error;
+    // }
+    const containerPrefix = `${process.env.CTF_LABEL}`;
+    return deleteByPrefix(containerPrefix);
 }
 
+/**
+ * Check if the flag is correct.
+ * @param {string} imageName - Docker image name.
+ * @param {string} userId - User Id.
+ * @param {string} flag - Flag to be checked.
+ * @returns {boolean} - Returns true if the flag is correct, false otherwise.
+ */
 async function checkFlag(imageName, userId, flag) {
     const containerName = `${userId}_${imageName}`;
     const labelName = `${process.env.CTF_LABEL}_${containerName}`;
@@ -154,36 +163,40 @@ async function deleteContainerByLabel(imageName, userId) {
  */
 async function deleteContainerByUser(userId) {
     const containerPrefix = `${process.env.CTF_LABEL}_${userId}`;
+    console.log(containerPrefix);
+    console.log('Deleting container');
     return deleteByPrefix(containerPrefix);
 }
 
 async function deleteByPrefix(containerPrefix) {
+    if (containerPrefix) {
+        try {
+            const containers = await docker.listContainers({ all: true });
 
-    try {
-        const containers = await docker.listContainers({ all: true });
+            const userContainers = containers.filter(container =>
+                container.Names.some(name => name.includes(containerPrefix))
+            );
 
-        const userContainers = containers.filter(container =>
-            container.Names.some(name => name.includes(containerPrefix))
-        );
+            if (userContainers.length === 0) {
+                console.log(`No containers found for prefix: ${prefix}`);
+                return true;
+            }
 
-        if (userContainers.length === 0) {
-            console.log(`No containers found for prefix: ${prefix}`);
-            return;
+            for (const containerInfo of userContainers.reverse()) {
+                const container = docker.getContainer(containerInfo.Id);
+
+                await container.kill();
+                await container.remove();
+                console.log(`Container ${containerInfo.Names[0]} deleted successfully.`);
+            }
+            return true;
+
+        } catch (error) {
+            console.error('Error deleting containers:', error);
+            throw error;
         }
 
-        for (const containerInfo of userContainers) {
-            const container = docker.getContainer(containerInfo.Id);
-
-            await container.kill();
-            await container.remove();
-            console.log(`Container ${containerInfo.Names[0]} deleted successfully.`);
-        }
-        return true;
-
-    } catch (error) {
-        console.error('Error deleting containers:', error);
-        throw error;
     }
 }
 
-export { createContainer, deleteContainerByUser, deleteContainerByLabel, deleteAllContainers, checkFlag };
+export { createContainer, deleteContainerByUser, deleteContainerByLabel, checkFlag, deleteAllContainers };
